Simplified delta computation; use .0 on floating-point constants
git-svn-id: https://svn.blender.org/svnroot/verse/trunk/verse@2658 49c17534-1d02-dc11-b283-0030488c597c

+++ v_network_out_que.c
@@ -224,16 +224,14 @@ boolean v_noq_send_queue(VNetOutQueue *queue, void *address)
 	uint32 seconds, fractions;
 	double delta;
 	
-	
 	data = queue->packet_buffer;
 	v_n_get_current_time(&seconds, &fractions);
-	delta = (double)(seconds - queue->seconds) + ((double)fractions - (double)queue->fractions) / (double) 0xffffffff;
+	delta = seconds - queue->seconds + (fractions - queue->fractions) / (double) 0xffffffff;
 	
 	if(queue->unsorted != NULL)	
 		v_noq_sort_unsorted(queue);
 
-
-	if(queue->unsent_size == 0 && delta < 1 && (queue->ack_nak == NULL || queue->ack_nak->next == NULL))
+	if(queue->unsent_size == 0 && delta < 1.0 && (queue->ack_nak == NULL || queue->ack_nak->next == NULL))
 		return TRUE;
 
 	if(delta > 2.0 && queue->unsent_size == 0 && queue->ack_nak == NULL)
